+++ src/adapters/wpapi.js
@@ -194,19 +194,23 @@ export default class WPAPIAdapter {
    * @return {Object}         The Request URL
    */
   getUrl({ wpRequest }) {
-    return wpRequest._renderURI();
+    return wpRequest.toString();
   }
 
   /**
    * Get aggregator for URL
    *
-   * Infers the aggregator identifier of a given URL to which all resulting resources are going to
-   * be associated with. An aggregator is a set containing resources indexed by its ids and by the
+   * Infers the aggregator identifier of a given URL to which a resulting resource will be
+   * associated with. `additionalData` is available so the decision might also be based on the query
+   * or on the own resource.
+   *
+   * An aggregator is a set containing resources indexed by its ids and by the
    * its custom indexers.
    *
-   * @param  {String}      url          URL from which the aggregator will be infered
-   * @return {String|null} aggregatorID String to which all URL direct resources will be associated
-   *                                    with or null, if resources musn't be indexed
+   * @param  {String}      url            URL from which the aggregator will be infered
+   * @param  {Object|null} additionalData Available data about the expected resource
+   * @return {String|null} aggregatorID   String to which all URL direct resources will be
+   *                                      associated with or null, if resources musn't be indexed
    */
   getAggregator(url) {
     let uri = url.replace(this.api._options.endpoint, '').replace(/\?.*$/, '');