+++ test/mocks/actions/successfulSearchRequest.js
@@ -4,13 +4,13 @@ import collectionResponse from '../data/collectionResponse';
 export default {
   type: REDUX_WP_API_SUCCESS,
   payload: {
-    cacheID: '/namespace/any',
+    cacheID: '/namespace/search',
     page: 1,
     response: collectionResponse,
   },
   meta: {
     name: 'test',
-    url: 'http://wordpress.dev/wp-json/namespace/any',
+    url: 'http://wordpress.dev/wp-json/namespace/search',
     requestAt: Date.now(),
     responseAt: Date.now(),
     operation: 'get',