+++ test/mocks/actions/unsuccessfulModifyingRequest.js
@@ -6,7 +6,7 @@ export default {
   error: new Error('Network Failure'),
   meta: {
     name: 'test',
-    aggregator: 'any',
+    url: 'http://wordpress.dev/wp-json/namespace/any',
     // operation will be injected;
     requestAt: Date.now(),
     responseAt: Date.now(),