+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Some guidelines in reading this document:
 * Being that these are the early days of the repository, we have some code changes that were added directly and without much detail, for these we have a link to the commit instead of the PR.
 * Annotations starting with **[BC]** indicates breaking change.
 
+## [new release]
+* ([#22](https://github.com/log-oscon/redux-wpapi/pull/22))
+    * Adapter's `getAggregator` is now applied per resource and has `additionalData` in order decide, which might be either the query of the request or the resource itself.
+    * `getAggregator` setting is now available, so consumer might also decide per resource. As its third param, the `suggestedAggregator` is supplied contain the adapter's `getAggregator` return.
+
 ## 1.2.1
 * ([#21](https://github.com/log-oscon/redux-wpapi/pull/21))
     * Mark embedded resource as `partial` until they are fetched individually (closes [#20](https://github.com/log-oscon/redux-wpapi/issues/20))